requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
bcrypt==4.1.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.8.0
dj-database-url==0.5.0
Django==4.2.22
django-bootstrap3
django-gmailapi-backend==0.3.2
django-mail-templated==2.6.5
django-utils-six==2.0
django-widget-tweaks==1.5.0
fonttools==4.58.1
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
gunicorn==23.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
invoke==2.2.0
lxml==5.3.1
mysql-connector-python==9.3.0
oscrypto==1.3.0
packaging==24.1
paramiko==3.4.0
pillow==10.4.0
proto-plus==1.26.1
# protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.3
pypdf==6.0.0
pyphen==0.17.2
pypng==0.20220715.0
python-bidi==0.6.6
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.4
rsa==4.9.1
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.5.0
urlquote==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
google-generativeai==0.8.3
